Fix daily repeat of last trigger
The current check in the binary sensor uses the date time attribute of 'seconds' however, this shows the number of seconds within a day since the event occurred and does not account for it being more than 24 hours. The proposed change adds the timeout to the last trigger and compares that to 'now' so that any time lapse will be accounted for.